alloc code done

+++ MP4/MP4_Sources/MP4_Sources/vm_pool.C
@@ -58,6 +58,8 @@ VMPool::VMPool(unsigned long  _base_address,
     //initialize arrays
     
     // setting all elements to 0
+
+    /*
     for (unsigned int i = 0; i < 512; i++){
         freelist_start_arr  [i] = NULL;
         freelist_end_arr    [i] = NULL;
@@ -65,7 +67,7 @@ VMPool::VMPool(unsigned long  _base_address,
         alloclist_end_arr   [i] = NULL;
     }
                 Console::puts("         -> cleaned lists\n");
-
+    */
 
     freelist_start_arr[0] = this->_base_address + (2*PAGE_SIZE);
     freelist_end_arr[0]   = this->_base_address + this->_size -1;
@@ -130,8 +132,8 @@ void VMPool::release(unsigned long _start_address) {
 
 bool VMPool::is_legitimate(unsigned long _address) {
 
-                Console::puts("         -> is_legitimate: base_address     =");Console::puti(this->_base_address);Console::puts("\n");
-                Console::puts("         -> is_legitimate: checking address =");Console::puti(_address);Console::puts("\n");
+                Console::puts("         -> is_legitimate: base_address     =");Console::puti((unsigned int)this->_base_address);Console::puts("\n");
+                Console::puts("         -> is_legitimate: checking address =");Console::puti((unsigned int)_address);Console::puts("\n");
 
     // if address belongs to free/alloc lists' region
     if((this->_base_address <= _address) && ((this->_base_address + 2*PAGE_SIZE) > _address)){
